refactor(core): move the collection/ inline test modules to sibling files (#1918)
First of four lots for velesdb-core, the last and largest crate of the
campaign. Twenty-eight files under collection/, split out by directory so
each PR stays reviewable.

Three shapes beyond the straight move, all found by auditing with the CI
guard's own block parser before touching anything:

- clustered_index.rs, search/query/distinct.rs, search/vector.rs and
  streaming/delta.rs already have a <stem>_tests.rs sibling holding
  separate integration tests, so their inline unit tests land in
  <stem>_unit_tests.rs and the existing files are left alone.
- collection_config.rs and config_serde.rs name their module something
  other than tests (rescore_config_tests, config_serde_tests); the name
  carries information, so it is kept and the sibling is named after it.

The sorted test-name list is byte-identical before and after, re-checked
after rustfmt: 5357 tests, 5341 passing and 16 pre-existing ignored.

The inline-tests baseline shrinks by the twenty-eight collection/ lines,
leaving 86 files.
