+++ scripts/inline-tests-baseline.txt
@@ -5,34 +5,6 @@ crates/tauri-plugin-velesdb/src/helpers.rs	200
 crates/tauri-plugin-velesdb/src/lib.rs	36
 crates/velesdb-core/src/agent/memory_helpers.rs	160
 crates/velesdb-core/src/api_types/requests.rs	65
-crates/velesdb-core/src/collection/collection_config.rs	104
-crates/velesdb-core/src/collection/config_serde.rs	37
-crates/velesdb-core/src/collection/core/sparse_lifecycle.rs	25
-crates/velesdb-core/src/collection/core/statistics.rs	103
-crates/velesdb-core/src/collection/graph/clustered_index.rs	117
-crates/velesdb-core/src/collection/graph/helpers.rs	71
-crates/velesdb-core/src/collection/graph_collection.rs	311
-crates/velesdb-core/src/collection/payload_size.rs	47
-crates/velesdb-core/src/collection/query_cost/cost_factors.rs	66
-crates/velesdb-core/src/collection/query_cost/cost_model.rs	91
-crates/velesdb-core/src/collection/query_cost/feedback.rs	117
-crates/velesdb-core/src/collection/query_cost/query_executor.rs	173
-crates/velesdb-core/src/collection/search/query/bounded_top_k.rs	67
-crates/velesdb-core/src/collection/search/query/condition_tree.rs	132
-crates/velesdb-core/src/collection/search/query/distinct.rs	129
-crates/velesdb-core/src/collection/search/query/fused_dispatch.rs	69
-crates/velesdb-core/src/collection/search/query/match_dispatch.rs	362
-crates/velesdb-core/src/collection/search/query/match_exec/index_prefilter.rs	222
-crates/velesdb-core/src/collection/search/query/match_exec/vector_first.rs	151
-crates/velesdb-core/src/collection/search/query/projection.rs	265
-crates/velesdb-core/src/collection/search/query/set_operations.rs	150
-crates/velesdb-core/src/collection/search/query/similarity_filter.rs	24
-crates/velesdb-core/src/collection/search/query/validation.rs	148
-crates/velesdb-core/src/collection/search/query/vector_group_by.rs	194
-crates/velesdb-core/src/collection/search/query/where_eval.rs	101
-crates/velesdb-core/src/collection/search/resolve.rs	155
-crates/velesdb-core/src/collection/search/vector.rs	42
-crates/velesdb-core/src/collection/streaming/delta.rs	290
 crates/velesdb-core/src/config.rs	103
 crates/velesdb-core/src/distance.rs	35
 crates/velesdb-core/src/fault_injection.rs	48
